requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.11.6
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.2.post1
async-timeout==4.0.3
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
coredis==4.17.0
Deprecated==1.2.15
distro==1.9.0
docstring_parser==0.16
fhaviary @ git+https://github.com/Future-House/aviary@21d822f32b02a3d2192826e9fecd4c8f7ba37f12
filelock==3.16.1
frozenlist==1.5.0
fsspec==2024.10.0
h11==0.14.0
html2text==2024.2.26
httpcore==1.0.7
httpx==0.27.2
huggingface-hub==0.26.2
idna==3.10
importlib_metadata==8.5.0
importlib_resources==6.4.5
Jinja2==3.1.4
jiter==0.7.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
latexcodec==3.0.0
limits==3.13.0
litellm==1.52.12
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
multidict==6.1.0
numpy==2.1.3
openai==1.55.0
packaging==24.2
paper-qa @ git+https://github.com/Future-House/paper-qa@cef9e2eef1c1c8b6470b7523958d2dbb282c85c0
propcache==0.2.0
pybtex==0.24.0
pydantic==2.10.0
pydantic-settings==2.6.1
pydantic_core==2.27.0
Pygments==2.18.0
Pympler==1.1
PyMuPDF==1.24.14
python-dotenv==1.0.1
PyYAML==6.0.2
redis==5.2.0
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
rich==13.9.4
rpds-py==0.21.0
setuptools==75.6.0
six==1.16.0
sniffio==1.3.1
tantivy==0.22.0
tenacity==9.0.0
tiktoken==0.8.0
tokenizers==0.20.3
tqdm==4.67.0
typing_extensions==4.12.2
urllib3==2.2.3
wrapt==1.16.0
yarl==1.17.2
zipp==3.21.0
